refactor(Table): 解决初始排序错误问题&优化排序代码
