+++ packages/devui-vue/devui/table/src/components/header-th/use-header-th.ts
@@ -1,4 +1,4 @@
-import { ref, Ref, computed, getCurrentInstance, inject, onMounted } from 'vue';
+import { ref, Ref, computed, getCurrentInstance, inject, onMounted, nextTick } from 'vue';
 import { Column, FilterConfig, SortDirection } from '../column/column-types';
 import { TABLE_TOKEN, ITableInstanceAndDefaultRow } from '../../table-types';
 import { UseSort, UseFilter, UseBaseRender, UseDragColumnWidth } from './header-th-types';
@@ -15,13 +15,14 @@ export function useBaseRender(column: Ref<Column>): UseBaseRender {
 export function useSort(column: Ref<Column>): UseSort {
   const table = inject(TABLE_TOKEN) as ITableInstanceAndDefaultRow;
   const store = table.store;
-  const direction = ref<SortDirection | undefined>(column.value.sortDirection);
+  const direction = ref<SortDirection>(column.value.sortDirection || '');
   const sortClass = computed(() => ({
     'sort-active': Boolean(direction.value),
   }));
   const thInstance = getCurrentInstance();
-  thInstance && store.states.thList.push(thInstance);
-  onMounted(() => {
+  thInstance && store.collectTh(thInstance);
+  onMounted(async () => {
+    await nextTick();
     column.value.sortable && column.value.sortDirection && store.sortData?.(direction.value, column.value.sortMethod);
   });
   const execClearSortOrder = () => {
